chore(contracts): update parameter type to int32 due to it being a small number

+++ contracts/extension/GovernorVotesSuperQuorumFraction.sol
@@ -22,7 +22,7 @@ abstract contract GovernorVotesSuperQuorumFraction is GovernorVotes {
 
     /// @notice Initializes the contract with a super quorum numerator.
     /// @param quorumNumeratorValue The initial super quorum numerator.
-    constructor(uint256 quorumNumeratorValue) {
+    constructor(uint32 quorumNumeratorValue) {
         _updateSuperQuorumNumerator(quorumNumeratorValue);
     }
